Release/0.8.5 rc0

+++ core/store/models/service_agreement.go
@@ -20,7 +20,7 @@ import (
 // Encumbrance connects job specifications with on-chain encumbrances.
 type Encumbrance struct {
 	// Corresponds to requestDigest in solidity ServiceAgreement struct
-	ID uint `json:"-" gorm:"primary_key;auto_increment"`
+	ID int64 `json:"-" gorm:"primary_key;auto_increment"`
 	// Price to request a report based on this agreement
 	Payment *assets.Link `json:"payment,omitempty"`
 	// Expiration is the amount of time an oracle has to answer a request
@@ -52,7 +52,7 @@ type ServiceAgreement struct {
 	ID            string      `json:"id" gorm:"primary_key"`
 	CreatedAt     time.Time   `json:"createdAt" gorm:"index"`
 	Encumbrance   Encumbrance `json:"encumbrance"`
-	EncumbranceID uint        `json:"-"`
+	EncumbranceID int64       `json:"-"`
 	RequestBody   string      `json:"requestBody"`
 	Signature     Signature   `json:"signature" gorm:"type:varchar(255)"`
 	JobSpec       JobSpec     `gorm:"foreignkey:JobSpecID"`
